+++ include/config.h
@@ -45,12 +45,16 @@
 #define ESS_PLATFORM_LINUX   			0/** @brief If defined compiled backend for generic linux (openal) */
 #define ESS_PLATFORM_WINDOWS  0/** @brief If defined compiled backend for  generic windows (openal, wasapi) */
 
-#define ESS_DEFAULT_SERVER_PORT 8082
+#define ESS_DEFAULT_SERVER_PORT "8082"
 #define ESS_DEFAULT_SERVER_PROTOCOL ESS_PROTOCOL_UDP
-#define ESS_DEFAULT_SERVER_FAMILY ESS_FAMILY_IP4
+#define ESS_DEFAULT_SERVER_FAMILY ESS_FAMILY_IP6
 // on IP4 : 0.0.0.0
-#define ESS_DEFAULT_SERVER_HOST "::"
 
+#if ESS_DEFAULT_SERVER_FAMILY == ESS_FAMILY_IP4
+#define ESS_DEFAULT_SERVER_HOST "0.0.0.0"
+#else
+#define ESS_DEFAULT_SERVER_HOST "::"
+#endif
 
 #if  ESS_PLATFORM_ESP32 == 1
 	#define ESS_CONFIG_SEMAPHORE_GENERIC